Add solvation function (#137)
Add solvation function as a method for Configuration, which allows the configuration to be solvated with a solvent of choice. The solvent can be selected from a database of solvents stored in autode with a corresponding density provided in a database in mlptrain. Alternatively, one can also specify their own solvent molecule and desired density. PR also includes tests for the function.